fix: Daphne hang
